# docs/openapi-v2.yaml
openapi: "3.0.3"
info:
  title: "Free Trial API — V2"
  version: "1.0.0"
  description: |
    Free Trial API for testers. Get started by creating an API key, then use it to manage users.

    **Base URL:** `http://85.193.81.51:3001`
servers:
  - url: "http://85.193.81.51:3001"
    description: "Production"
paths:
  /free/v2/api/keys:
    post:
      summary: "Create API key"
      description: "Creates a new API key valid for 24 hours."
      operationId: "createApiKeyV2"
      tags: ["Keys"]
      responses:
        "201":
          description: "API key created"
          content:
            application/json:
              schema:
                type: object
                required: [key, expires_at]
                properties:
                  key:
                    type: string
                    example: "free-trial-1a2b3c4d"
                  expires_at:
                    type: string
                    format: "date-time"
                    example: "2026-05-25T12:00:00.000Z"

  /free/v2/api/users:
    post:
      summary: "Create user"
      description: "Creates a new user with a name and age."
      operationId: "createUserV2"
      tags: ["Users"]
      parameters:
        - name: "X-Fix-Bug"
          in: "header"
          required: true
          schema:
            type: string
          description: "Your API key"
          example: "free-trial-1a2b3c4d"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, age]
              properties:
                name:
                  type: string
                  description: "User's name (non-empty, trimmed)"
                  example: "Alice"
                age:
                  type: integer
                  description: "User's age (18–65 inclusive)"
                  minimum: 18
                  maximum: 65
                  example: 25
      responses:
        "201":
          description: "User created"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    required: [id, name, age, created_at]
                    properties:
                      id:
                        type: integer
                        example: 42
                      name:
                        type: string
                        example: "Alice"
                      age:
                        type: integer
                        example: 25
                      created_at:
                        type: string
                        format: "date-time"
        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or invalid API key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: "List users"
      description: "Returns a list of all users. Supports optional limit."
      operationId: "listUsersV2"
      tags: ["Users"]
      parameters:
        - name: "X-Fix-Bug"
          in: "header"
          required: true
          schema:
            type: string
          description: "Your API key"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: integer
            minimum: 1
          description: "Maximum number of users to return"
          example: 10
      responses:
        "200":
          description: "List of users"
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserSummary"
        "401":
          description: "Missing or invalid API key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /free/v2/api/users/{id}:
    get:
      summary: "Get user by ID"
      description: "Returns a single user by their ID."
      operationId: "getUserByIdV2"
      tags: ["Users"]
      parameters:
        - name: "X-Fix-Bug"
          in: "header"
          required: true
          schema:
            type: string
          description: "Your API key"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: integer
            minimum: 1
          description: "User ID"
          example: 5
      responses:
        "200":
          description: "User found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserSummary"
        "400":
          description: "Invalid ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or invalid API key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: "Update user"
      description: "Updates an existing user's name and/or age."
      operationId: "updateUserV2"
      tags: ["Users"]
      parameters:
        - name: "X-Fix-Bug"
          in: "header"
          required: true
          schema:
            type: string
          description: "Your API key"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: integer
            minimum: 1
          description: "User ID"
          example: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "New name (non-empty, trimmed)"
                  example: "Alice Updated"
                age:
                  type: integer
                  description: "New age (18–65 inclusive)"
                  minimum: 18
                  maximum: 65
                  example: 30
      responses:
        "200":
          description: "User updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/UserSummary"
        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or invalid API key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: "Delete user"
      description: "Deletes a user by their ID."
      operationId: "deleteUserV2"
      tags: ["Users"]
      parameters:
        - name: "X-Fix-Bug"
          in: "header"
          required: true
          schema:
            type: string
          description: "Your API key"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: integer
            minimum: 1
          description: "User ID"
          example: 5
      responses:
        "200":
          description: "User deleted"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Deleted 1 user"
                  user:
                    type: string
                    example: "Alice"
        "400":
          description: "Invalid ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or invalid API key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /ping:
    get:
      summary: "Health check"
      description: "Returns service status and version."
      operationId: "healthCheck"
      tags: ["System"]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: "date-time"
                  version:
                    type: string
                    example: "v1+v2"

components:
  schemas:
    UserSummary:
      type: object
      required: [id, name, age]
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: "Alice"
        age:
          type: integer
          example: 25
        created_at:
          type: string
          format: "date-time"

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: "Error message"
        retry_after:
          type: integer
          description: "Seconds to wait before retrying (only for 429)"